pyproject.toml
[tool.poetry]
name = "shopify-admin-chatgpt-plugin"
version = "0.1.0"
description = ""
authors = ["Alex Meyer <acomeyer@gmail.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
shopifyapi = "^12.2.0"
requests = "^2.28.2"
pydantic = "^1.10.7"
sqlalchemy = "^2.0.9"
psycopg2 = "^2.9.6"

[tool.poetry.scripts]
start = "server.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
